requirements.txt
openai>=1.14.0
pyautogui>=0.9.54
Pillow>=10.0.0
pytesseract>=0.3.10
rich>=13.0.0
python-dotenv>=1.0.0
